backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-SQLAlchemy